requirements.txt
numpy
langchain
langchain-community
pypdf
faiss-cpu
PyPDF2
ollama
sentence-transformers
flask